entregando integração entre APIs

+++ Registration/Registration.Application/UseCases/DeletePetUseCase.cs
@@ -1,4 +1,5 @@
 ﻿using Registration.Domain.Interfaces.IRepositories;
+using Registration.Domain.Interfaces.IServices;
 using System;
 using System.Collections.Generic;
 using System.Text;
@@ -8,12 +9,14 @@ namespace Registration.Application.UseCases
     public class DeletePetUseCase
     {
         private readonly IPetRepository _petRepository;
-        public DeletePetUseCase(IPetRepository petRepository)
+        private readonly IAppointmentService _appointmentService;
+        public DeletePetUseCase(IPetRepository petRepository, IAppointmentService appointmentService)
         {
             _petRepository = petRepository;
+            _appointmentService = appointmentService;
         }
 
-        public bool Run(Guid id)
+        public async Task Run(Guid id)
         {
             var pet = _petRepository.GetById(id);
 
@@ -23,9 +26,13 @@ namespace Registration.Application.UseCases
             }
 
             //TODO: Adicionar validação para verificar se o pet pode ser deletado (ex: não pode ser deletado se tiver consultas agendadas)
+            var hasFuture = await _appointmentService.HasFutureAppointmentsAsync(id);
+
+            if (hasFuture)
+                throw new Exception("Pet cannot be deleted with future appointments.");
 
             _petRepository.Delete(id);
-            return true;
+            //return true;
         }
     }
 }
